# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Tripline API
  version: 0.1.0
  description: "Self-hosted dependency drift monitor. Authenticate with the session cookie (browser; state-changing requests need header `X-Tripline-CSRF: 1`) or a read-only bearer API token."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tripline_session
    token:
      type: http
      scheme: bearer
      description: Read-only API token (tl_…)
  schemas:
    def-0:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details: {}
      title: ErrorResponse
paths:
...
  /api/v1/monitors:
    get:
      summary: List monitors with recent check durations and open drift counts
      tags: &a4
        - monitors
      parameters:
        - schema:
            type: string
            maxLength: 200
          in: query
          name: q
          required: false
        - schema:
            enum:
              - http
              - llm
              - mcp
          in: query
          name: kind
          required: false
        - schema:
            enum:
              - unknown
              - up
              - down
          in: query
          name: status
          required: false
        - schema:
            type: string
            minLength: 1
            maxLength: 40
            pattern: ^[A-Za-z0-9_.-]+$
          in: query
          name: tag
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - monitors
                properties:
                  monitors:
                    type: array
                    items:
                      allOf:
                        - type: object
                          required:
                            - id
                            - name
                            - kind
                            - enabled
                            - intervalSeconds
                            - timeoutMs
                            - failureThreshold
                            - baselineSamples
                            - driftEnabled
                            - ignorePaths
                            - tags
                            - config
                            - secretKeys
                            - status
                            - consecutiveFailures
                            - lastCheckedAt
                            - lastDurationMs
                            - lastError
                            - nextRunAt
                            - createdBy
                            - createdAt
                            - updatedAt
                          properties:
                            id:
                              type: string
                              format: uuid
                            name:
                              type: string
                            kind:
                              enum:
                                - http
                                - llm
                                - mcp
                            enabled:
                              type: boolean
                            intervalSeconds:
                              type: integer
                            timeoutMs:
                              type: integer
                            failureThreshold:
                              type: integer
                            baselineSamples:
                              type: integer
                            driftEnabled:
                              type: boolean
                            ignorePaths:
                              type: array
                              items:
                                type: string
                            tags:
                              type: array
                              items:
                                type: string
                            config:
                              type: object
                              additionalProperties: {}
                            secretKeys:
                              type: object
                              required:
                                - headers
                                - apiKey
                              properties:
                                headers:
                                  type: array
                                  items:
                                    type: string
                                apiKey:
                                  type: boolean
                            status:
                              enum:
                                - unknown
                                - up
                                - down
                            consecutiveFailures:
                              type: integer
                            lastCheckedAt:
                              anyOf:
                                - type: string
                                  format: date-time
                                - type: "null"
                            lastDurationMs:
                              anyOf:
                                - type: integer
                                - type: "null"
                            lastError:
                              anyOf:
                                - type: string
                                - type: "null"
                            nextRunAt:
                              type: string
                              format: date-time
                            createdBy:
                              anyOf:
                                - type: string
                                  format: uuid
                                - type: "null"
                            createdAt:
                              type: string
                              format: date-time
                            updatedAt:
                              type: string
                              format: date-time
                        - type: object
                          required:
                            - recent
                            - openDrift
                          properties:
                            recent:
                              type: array
                              items:
                                type: object
                                required:
                                  - d
                                  - ok
                                properties:
                                  d:
                                    type: integer
                                  ok:
                                    type: boolean
                            openDrift:
                              type: object
                              required:
                                - open
                                - worst
                              properties:
                                open:
                                  type: integer
                                worst:
                                  anyOf:
                                    - enum:
                                        - breaking
                                        - warning
                                        - info
                                    - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    post:
      summary: Create a monitor
      tags: *a4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - config
                - kind
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                enabled:
                  type: boolean
                intervalSeconds:
                  type: integer
                  minimum: 30
                  maximum: 86400
                timeoutMs:
                  type: integer
                  minimum: 500
                  maximum: 60000
                failureThreshold:
                  type: integer
                  minimum: 1
                  maximum: 20
                baselineSamples:
                  type: integer
                  minimum: 1
                  maximum: 20
                driftEnabled:
                  type: boolean
                ignorePaths:
                  type: array
                  items:
                    type: string
                    minLength: 1
                    maxLength: 512
                  maxItems: 50
                tags:
                  type: array
                  items:
                    type: string
                    minLength: 1
                    maxLength: 40
                    pattern: ^[A-Za-z0-9_.-]+$
                  maxItems: 20
                config:
                  type: object
                  description: Kind-specific configuration; see API_DOCUMENTATION.md.
                  additionalProperties: {}
                secrets:
                  type: object
                  properties:
                    headers:
                      type: object
                      additionalProperties:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 8192
                          - type: "null"
                    apiKey:
                      anyOf:
                        - type: string
                          minLength: 1
                          maxLength: 4096
                        - type: "null"
                  additionalProperties: false
                  description: Secret values are write-only. `null` removes a secret; omitting keeps it.
                kind:
                  enum:
                    - http
                    - llm
                    - mcp
              additionalProperties: false
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - monitor
                properties:
                  monitor:
                    type: object
                    required:
                      - id
                      - name
                      - kind
                      - enabled
                      - intervalSeconds
                      - timeoutMs
                      - failureThreshold
                      - baselineSamples
                      - driftEnabled
                      - ignorePaths
                      - tags
                      - config
                      - secretKeys
                      - status
                      - consecutiveFailures
                      - lastCheckedAt
                      - lastDurationMs
                      - lastError
                      - nextRunAt
                      - createdBy
                      - createdAt
                      - updatedAt
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      kind:
                        enum:
                          - http
                          - llm
                          - mcp
                      enabled:
                        type: boolean
                      intervalSeconds:
                        type: integer
                      timeoutMs:
                        type: integer
                      failureThreshold:
                        type: integer
                      baselineSamples:
                        type: integer
                      driftEnabled:
                        type: boolean
                      ignorePaths:
                        type: array
                        items:
                          type: string
                      tags:
                        type: array
                        items:
                          type: string
                      config:
                        type: object
                        additionalProperties: {}
                      secretKeys:
                        type: object
                        required:
                          - headers
                          - apiKey
                        properties:
                          headers:
                            type: array
                            items:
                              type: string
                          apiKey:
                            type: boolean
                      status:
                        enum:
                          - unknown
                          - up
                          - down
                      consecutiveFailures:
                        type: integer
                      lastCheckedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
                      lastDurationMs:
                        anyOf:
                          - type: integer
                          - type: "null"
                      lastError:
                        anyOf:
                          - type: string
                          - type: "null"
                      nextRunAt:
                        type: string
                        format: date-time
                      createdBy:
                        anyOf:
                          - type: string
                            format: uuid
                          - type: "null"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/monitors/test:
    post:
      summary: Dry-run a monitor configuration without saving anything
      tags: *a4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kind
                - config
              properties:
                kind:
                  enum:
                    - http
                    - llm
                    - mcp
                config:
                  type: object
                  description: Kind-specific configuration; see API_DOCUMENTATION.md.
                  additionalProperties: {}
                secrets:
                  type: object
                  properties:
                    headers:
                      type: object
                      additionalProperties:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 8192
                          - type: "null"
                    apiKey:
                      anyOf:
                        - type: string
                          minLength: 1
                          maxLength: 4096
                        - type: "null"
                  additionalProperties: false
                  description: Secret values are write-only. `null` removes a secret; omitting keeps it.
                timeoutMs:
                  type: integer
                  minimum: 500
                  maximum: 60000
                monitorId:
                  type: string
                  format: uuid
                  description: Reuse the stored secrets of this monitor
              additionalProperties: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - durationMs
                  - message
                  - assertions
                  - signature
                  - meta
                  - observedPaths
                properties:
                  ok:
                    type: boolean
                  statusCode:
                    type: integer
                  durationMs:
                    type: integer
                  errorCode:
                    type: string
                  message:
                    type: string
                  assertions:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - ok
                        - message
                      properties:
                        name:
                          type: string
                        ok:
                          type: boolean
                        message:
                          type: string
                  signature:
                    type: object
                    additionalProperties:
                      type: object
                      required:
                        - value
                        - severity
                        - label
                      properties:
                        value:
                          type: string
                        severity:
                          enum:
                            - breaking
                            - warning
                            - info
                        label:
                          type: string
                  meta:
                    type: object
                    additionalProperties: {}
                  excerpt:
                    type: string
                  observedPaths:
                    type: integer
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...
  /api/v1/monitors/{id}/run:
    post:
      summary: Run a check now and record it
      tags: *a4
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - result
                  - status
                  - incident
                  - baseline
                  - drift
                properties:
                  result:
                    type: object
                    required:
                      - id
                      - monitorId
                      - startedAt
                      - durationMs
                      - ok
                      - statusCode
                      - errorCode
                      - message
                      - assertions
                      - meta
                      - responseExcerpt
                    properties:
                      id:
                        type: integer
                      monitorId:
                        type: string
                        format: uuid
                      startedAt:
                        type: string
                        format: date-time
                      durationMs:
                        type: integer
                      ok:
                        type: boolean
                      statusCode:
                        anyOf:
                          - type: integer
                          - type: "null"
                      errorCode:
                        anyOf:
                          - type: string
                          - type: "null"
                      message:
                        anyOf:
                          - type: string
                          - type: "null"
                      assertions:
                        type: array
                        items:
                          type: object
                          required:
                            - name
                            - ok
                            - message
                          properties:
                            name:
                              type: string
                            ok:
                              type: boolean
                            message:
                              type: string
                      meta:
                        type: object
                        additionalProperties: {}
                      responseExcerpt:
                        anyOf:
                          - type: string
                          - type: "null"
                  status:
                    enum:
                      - unknown
                      - up
                      - down
                  incident:
                    anyOf:
                      - enum:
                          - opened
                          - resolved
                          - ongoing
                      - type: "null"
                  baseline:
                    enum:
                      - learning
                      - locked
                      - disabled
                      - skipped
                  drift:
                    anyOf:
                      - type: object
                        required:
                          - eventId
                          - severity
                          - isNew
                          - changes
                        properties:
                          eventId:
                            type: string
                          severity:
                            enum:
                              - breaking
                              - warning
                              - info
                          isNew:
                            type: boolean
                          changes:
                            type: integer
                      - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...
servers:
  - url: /
security:
  - session: []
  - token: []